Add author creation functionality and refactor DTO packages

+++ src/main/resources/db/migration/V0_3_0__create_book_author_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE book_author
 (
-    book_id   INT NOT NULL,
-    author_id INT NOT NULL,
+    book_id   BIGINT NOT NULL,
+    author_id BIGINT NOT NULL,
 
     PRIMARY KEY (book_id, author_id),
     FOREIGN KEY (book_id) REFERENCES book (id) ON DELETE CASCADE,
